services/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pydantic==2.7.1
SQLAlchemy==2.0.32
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.2.2
jinja2==3.1.4
PyYAML==6.0.2
